Use the interfaces
