--- torch_spyre/_inductor/layout_backend.py
@@ -1,58 +1,0 @@
-# Copyright 2026 The Torch-Spyre Authors.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-from dataclasses import dataclass
-from typing import Optional
-from abc import ABC, abstractmethod
-
-
-@dataclass
-class LifetimeBoundBuffer:
-    """
-    Defines the data fields required for a layout solver.
-    The required heuristics are implementation defined.
-    """
-
-    name: str
-    size: int
-    start_time: int
-    end_time: int
-    heuristic: Optional[float] = None
-    address: Optional[int] = None
-
-
-class MemoryPlanSolver(ABC):
-    """
-    An abstract class for defining algorithms which solve
-    memory layout patterns based on provided sizes, lifetimes,
-    and optional heuristics based on the implementation details.
-    """
-
-    @abstractmethod
-    def plan_layout(
-        self, buffers: list[LifetimeBoundBuffer]
-    ) -> list[LifetimeBoundBuffer]:
-        """
-        Utilizes an implementation defined algorithm to determine
-        if and where buffers should be placed in scratchpad memory based
-        on their attributes.
-
-        Args:
-            buffers (list[LifetimeBoundBuffer]): The set of candidate buffers for memory planning
-
-        Returns:
-            list[LifetimeBoundBuffer]: The set of buffers with their placements defined.
-        """
-        pass
