+++ torch_spyre/_inductor/scratchpad/allocator.py
@@ -13,94 +13,45 @@
 # limitations under the License.
 
 from abc import ABC, abstractmethod
-from typing import Optional
+import math
+from typing import Callable, Any, Optional
+
+from torch._inductor.ir import (
+    ComputedBuffer,
+    MutationLayoutSHOULDREMOVE,
+)
+from torch._inductor.lowering import lowerings, clone as clone_lowering
+from torch._inductor.ops_handler import WrapperHandler
+from torch._inductor.virtualized import V
 from torch._inductor.graph import GraphLowering
+from .ir import FixedTiledLayout, TensorBox
+from . import config
+
 from torch_spyre._inductor.scratchpad.plan_solver import (
-    MemoryPlanSolver,
+    GreedyLayoutSolver,
     LifetimeBoundBuffer,
+    MemoryPlanSolver,
 )
 from torch_spyre._inductor.scratchpad.passes import ScratchpadOptimizationPass
 from torch_spyre._inductor.scratchpad.utility import (
-    is_permissible_op,
-    is_core_division_equal,
-    is_graph_edge,
-    calculate_liveness,
-    push_allocation,
-    mem_usage_by_buffer,
+    get_buffer_users,
+    get_ncores_for_buffers,
 )
 
 
-class GraphBufferConverstion:
-    def __init__(
-        self,
-        permitted_ops: list[str] = [
-            "max",
-            "sum",
-            "exp",
-            "sub",
-        ],
-        in_place_ops: list[str] = ["exp", "sub"],
-    ):
-        self.permitted_ops = permitted_ops
-        self.in_place_ops = in_place_ops
+OP_OUTPUT_GOOD_FOR_LX_REUSE = [
+    "max",
+    "sum",
+    # "clone",
+    "exp",
+    "sub",
+    # "mul",
+]
 
-    def _filter_buffers(
-        self, graph: GraphLowering, buffers: list[LifetimeBoundBuffer]
-    ) -> list[LifetimeBoundBuffer]:
-        permissible_ops = is_permissible_op(graph, self.permitted_ops)
-        core_division_match = is_core_division_equal(graph)
-        input_output = is_graph_edge(graph)
-
-        drop_list = (
-            [key for key, permissible in permissible_ops.items() if not permissible] + 
-            [key for key, permissible in core_division_match.items() if not permissible] +
-            [key for key, edge in input_output.items() if edge]
-        )
-
-        return [b for b in buffers if b.name not in drop_list]
-
-    def _build_bound_buffers(
-        self,
-        graph: GraphLowering,
-        heuristics: dict[str, float] = {},
-        in_place: dict[str, list[str]] = {},
-    ) -> list[LifetimeBoundBuffer]:
-        lifetime_starts, lifetime_ends = calculate_liveness(graph)
-        sizes = mem_usage_by_buffer(graph)
-
-        assert lifetime_starts.keys() == sizes.keys(), (
-            "The keys in lifetimes and sizes must match"
-        )
-
-        assert lifetime_starts.keys() == lifetime_ends.keys(), (
-            "The keys in lifetimes must match"
-        )
-
-        return [
-            LifetimeBoundBuffer(
-                buffer_name,
-                sizes[buffer_name],
-                lifetime_starts[buffer_name],
-                lifetime_ends[buffer_name],
-                heuristics[buffer_name] if buffer_name in heuristics else None,
-                in_place=in_place[buffer_name] if buffer_name in in_place else [],
-            )
-            for buffer_name in lifetime_starts.keys()
-        ]
-
-    def _determine_in_place(
-        self, graph: GraphLowering, in_place_ops: list[str]
-    ) -> dict[str, list[str]]:
-        return {}
-
-    def generate_buffers(self, graph: GraphLowering) -> list[LifetimeBoundBuffer]:
-        heuristics: dict[
-            str, float
-        ] = {}  # TODO: implement a concrete way of calculating this from a graph
-        in_place = self._determine_in_place(graph, self.in_place_ops)
-        buffers = self._build_bound_buffers(graph, heuristics, in_place)
-        filtered_buffers = self._filter_buffers(graph, buffers)
-        return filtered_buffers
+OP_GOOD_FOR_LX_INPLACE = [
+    "exp",
+    "sub",
+]
 
 
 class ScratchpadAllocator(ABC):
@@ -119,28 +70,294 @@ class ScratchpadAllocator(ABC):
         """
         pass
 
+    def op_output_good_for_lx_reuse(self, op: Any) -> bool:
+        return (
+            isinstance(op, ComputedBuffer)
+            and not isinstance(op.layout, MutationLayoutSHOULDREMOVE)
+            and (
+                config.allow_all_ops_in_lx_planning
+                or op._opname in OP_OUTPUT_GOOD_FOR_LX_REUSE
+            )
+        )
+
+    def op_good_for_lx_inplace(self, org_op_name: str) -> bool:
+        return any(op in org_op_name for op in OP_GOOD_FOR_LX_INPLACE)
+
+    def generate_buffers(self, graph: GraphLowering) -> list[LifetimeBoundBuffer]:
+        in_place = self._determine_in_place(graph)
+        buffers = self._build_bound_buffers(graph, in_place)
+        filtered_buffers = self._filter_buffers(graph, buffers)
+        return filtered_buffers
+
+    def _filter_buffers(
+        self, graph: GraphLowering, buffers: list[LifetimeBoundBuffer]
+    ) -> list[LifetimeBoundBuffer]:
+        """
+        From the list of buffers, drop buffers that are outputs of
+        unpermitted ops, graph outputs, and graph inputs
+        """
+
+        drop_list = set()
+        for op in graph.operations:
+            if not self.op_output_good_for_lx_reuse(op.origin_node.target):
+                rw = op.get_read_writes()
+                for mem_dep in rw.writes:
+                    drop_list.add(mem_dep.name)
+
+        drop_list.update(graph.get_output_names())
+        drop_list.update(graph.graph_input_names)
+
+        return [b for b in buffers if b.name not in drop_list]
+
+    def _mem_usage_by_buffer(
+        self, graph: GraphLowering
+    ) -> dict[str, dict[str, bool | int]]:
+        """
+        Get a summary of memory usage for all operations
+        Detailed info of individual buf, e.g. mem_usage[<buf_name>], which has
+            "size", "core_div_mismatch", "size_per_core", "liveness_start",
+            "liveness_end" fields
+        """
+        mem_usage = {}
+        for buf_name, num_cores in get_ncores_for_buffers(graph).items():
+            buf = graph.get_buffer(buf_name)
+            dev_layout = buf.layout.device_layout
+            dev_size = (
+                math.prod(dev_layout.device_size[:-1]) * 128
+            )  # num_sticks * bytes_per_stick
+            mem_usage[buf_name] = {
+                "size": dev_size,
+                "size_per_core": dev_size // num_cores,
+                "core_div_mismatch": num_cores == -1,
+            }
+        self._calculate_liveness(graph, mem_usage)
+        return mem_usage
+
+    def _calculate_liveness(
+        self, graph: GraphLowering, mem_usage: dict[str, dict[str, bool | int]]
+    ) -> None:
+        for i, op in enumerate(graph.operations):
+            rw = op.get_read_writes()
+            for mem_dep in rw.reads | rw.writes:
+                buf_name = mem_dep.name
+                if "liveness_start" not in mem_usage[buf_name]:
+                    mem_usage[buf_name]["liveness_start"] = i
+                mem_usage[buf_name]["liveness_end"] = i + 1
+
+    def _build_bound_buffers(
+        self,
+        graph: GraphLowering,
+        in_place: dict[str, list[str]] = {},
+    ) -> list[LifetimeBoundBuffer]:
+        mem_usage = self._mem_usage_by_buffer(graph)
+
+        return [
+            LifetimeBoundBuffer(
+                buffer_name,
+                info["size_per_core"],
+                info["liveness_start"],
+                info["liveness_end"],
+                in_place=in_place[buffer_name] if buffer_name in in_place else [],
+            )
+            for buffer_name, info in mem_usage.items()
+            if not info["core_div_mismatch"]
+        ]
+
+    def _determine_in_place(
+        self,
+        graph: GraphLowering,
+    ) -> dict[str, list[str]]:
+        return {}
+
+    class NameSwapHandler(WrapperHandler):
+        def __init__(self, inner, name_map: dict[str, str]):
+            super().__init__(inner)
+            self._name_map = name_map
+
+        def load(self, name, index):
+            return super().load(self._name_map.get(name, name), index)
+
+    def create_Loop_hack_inner_fn(self, old_Loop, name_map):
+        """Use ops_handler to swap the name of buffers"""
+
+        def new_inner_fn(*args):
+            # Pointwise has 1 pos arg index while Reduction has 2, i.e. (index, rindex)
+            with V.set_ops_handler(self.NameSwapHandler(V.ops, name_map)):
+                return old_Loop.inner_fn(*args)
+
+        # old_Loop could be a Pointwise or Reduction.
+        kwargs = {k: getattr(old_Loop, k) for k in old_Loop.__dataclass_fields__.keys()}
+        kwargs["inner_fn"] = new_inner_fn
+        new_Loop = old_Loop.__class__(**kwargs)
+        # Additional attr that are not included in dataclass_fields. NOTE it relies on a
+        # special method to force reset attrs of a frozen dataclass, see ir.Loops.create()
+        new_Loop._post_init_setattr("origins", old_Loop.origins)
+        new_Loop._post_init_setattr("origin_node", old_Loop.origin_node)
+        new_Loop._post_init_setattr("traceback", old_Loop.traceback)
+        # .get_stack_traces() get info from "origins", no need to manually set anything
+        # LoopBody will be created later when we call CompBuf.recompute()
+
+        return new_Loop
+
+    def try_insert_clone_op_for_inputs(
+        self,
+        graph: GraphLowering,
+        lx_free_total: int,
+    ) -> None:
+        """
+        Check if any input tensors can fit onto scratchpad and needed more than once =>
+        Add corresponding "clone operation" to copy it to scratchpad and reduce HBM read.
+        """
+        buf_users = get_buffer_users(graph)
+        ncores = get_ncores_for_buffers(graph)
+        for inp_name in graph.graph_input_names:
+            buf = graph.get_buffer(inp_name)  # this is a TensorBox
+            dev_layout = buf.layout.device_layout
+            dev_size = math.prod(dev_layout.device_size[:-1]) * 128
+            is_on_lx = buf.layout.allocation != {}
+            used_only_once = len(buf_users[inp_name]) == 1
+            core_div_mismatch = ncores[inp_name] == -1
+            if (
+                used_only_once
+                or dev_size > lx_free_total
+                or is_on_lx
+                or core_div_mismatch
+            ):
+                continue
+
+            self.insert_op_after(buf, clone_lowering, buf_users, graph)
+            lx_free_total -= dev_size
+
+    def insert_op_after(
+        self,
+        buf: TensorBox,
+        lowering_func: Callable,
+        buf_users: dict,
+        graph: GraphLowering,
+    ) -> None:
+        """
+        Insert an operation using the provided lowering function (e.g. clone_lowering) in
+        GraphLowering.operations list after the given op (buf, a TensorBox representing a
+        ComputedBuffer). Will update GraphLowering FX graph and the operations list.
+        For example, original ops list looks like:
+            buf0 -> buf1 -> buf2
+        insert a clone of buf0 and let buf1 read from it, will become
+            buf0 ->(clone) buf3 -> buf1 ->buf2
+
+        NOTE:
+        - Simplified flow, everything is done before Scheduler. Only take care of FX and
+          GraphLowering. list operations will be updated inplace, no need to return.
+        - Even though it is not a necessary condition, we assume FX graph and Operations are
+          fully consistent and we will try to maintain it that way.
+        - To update existing users of the old buffer -> hack the inner_fn then refresh LoopIR
+        """
+        fx_graph = graph.graph
+
+        # Step 1: Add a new FX node for clone and update dependencies
+        buf_name = buf.data.data.name  # buf is a TensorBox
+        buf_fx = list(buf.origins)[0]  # .origin_node may not exist
+        old_users = list(buf_fx.users.keys())
+        # make sure the user-provided lowering_func is legit
+        assert lowering_func in lowerings.values(), (
+            f"The provided lowering function {lowering_func} is not properly registered."
+        )
+        LUTlower_func_to_op = {func: aten_op for aten_op, func in lowerings.items()}
+        user_aten_op = LUTlower_func_to_op[lowering_func]
+        # TODO this is a large dict, move it to upper scope so we only need to do it once
+        # aten_op is the overloaded version, e.g. ops.aten.clone.*out* instead of .default
+        fx_graph.inserting_after(buf_fx)
+        new_fx_node = fx_graph.create_node("call_function", user_aten_op, (buf_fx,))
+        for user in old_users:
+            user.args = tuple(new_fx_node if ar is buf_fx else ar for ar in user.args)
+        graph.orig_gm.recompile()
+
+        # Step 2: Create a new ComBuf of a Pointwise IR (need to support Reduction?)
+        pw_ir_tb = lowering_func(buf)  # a TensorBox wrapping a PointwiseIR
+        new_com_buf = ComputedBuffer(
+            name=None,
+            layout=FixedTiledLayout(
+                buf.layout.device,
+                buf.layout.dtype,
+                buf.layout.size,
+                buf.layout.stride,
+                buf.layout.device_layout,
+            ),  # create a new copy of FixedTiledLayout from buf's layout
+            data=pw_ir_tb.data.data,
+        )
+        new_com_buf.origins.add(new_fx_node)
+        new_com_buf.origin_node = new_fx_node
+        # TODO why arg0 ComputedBuffer doesn't have this attr?
+        new_com_buf.name = graph.register_buffer(new_com_buf)
+        graph.register_operation(new_com_buf)
+        new_buf_name = new_com_buf.name
+
+        # Step 3: Update graph_lowering.name_to_users (a list of TensorBox), eg, existing
+        # users of arg0, other than InpBuf and new_buf, should become users of new_buf.
+        users_of_inp, users_of_new_buf = [], []
+        for tb in graph.name_to_users[buf_name]:
+            if tb.data.data.name in [buf_name, new_buf_name]:
+                users_of_inp.append(tb)
+            else:
+                users_of_new_buf.append(tb)
+        graph.name_to_users[buf_name] = users_of_inp
+        graph.name_to_users[new_buf_name] = users_of_new_buf
+
+        # Step 4: Hack user nodes' inner_fn
+        for old_com_buf in buf_users[buf_name]:
+            # hack inner_fn with a nameSwapper ops handler and make a new LoopIR
+            new_Loop = self.create_Loop_hack_inner_fn(
+                old_com_buf.data, name_map={buf_name: new_buf_name}
+            )
+            old_com_buf.data = new_Loop
+
+        operations = graph.operations
+        # NOTE: operations is a reference to graph_lowering.operations, which is already
+        # updated when we call graph_lowering.register_operation() earlier. But the new Op
+        # was appended at the end of the list, need to insert at the correct position.
+        first_user = buf_users[buf_name][0]
+        idx_to_first_user = operations.index(first_user)
+        operations.remove(new_com_buf)
+        operations.insert(idx_to_first_user, new_com_buf)
+
+    def push_allocation(self, graph: GraphLowering, buffers: list[LifetimeBoundBuffer]):
+        # push the allocation into the code generation
+        for b in buffers:
+            if b.address is not None:
+                buf = graph.get_buffer(b.name)
+                layout = buf.get_layout()
+                layout.allocation["lx"] = b.address
+
 
 class DefaultAllocator(ScratchpadAllocator):
     def __init__(
         self,
-        layout_planning: MemoryPlanSolver,
+        layout_planning: Optional[MemoryPlanSolver] = None,
         pre_optimization_passes: list[ScratchpadOptimizationPass] = [],
         post_optimization_passes: list[ScratchpadOptimizationPass] = [],
-        buffer_source: Optional[GraphBufferConverstion] = None,
     ):
-        assert layout_planning is not None
+        if layout_planning is None:
+            layout_planning = GreedyLayoutSolver()
         self.pre_optimization_passes = pre_optimization_passes
         self.post_optimization_passes = post_optimization_passes
         self.layout_planning = layout_planning
-        self.buffer_source = (
-            buffer_source if buffer_source is not None else GraphBufferConverstion()
-        )
 
     def plan_allocation(self, graph: GraphLowering):
         for p in self.pre_optimization_passes:
             p.apply_pass(graph)
-        buffers = self.buffer_source.generate_buffers(graph)
+        buffers = self.generate_buffers(graph)
         allocation = self.layout_planning.plan_layout(buffers)
-        push_allocation(graph, allocation)
+        self.push_allocation(graph, allocation)
         for p in self.post_optimization_passes:
             p.apply_pass(graph)
+
+
+def scratchpad_planning(
+    graph: GraphLowering,
+    strategy: Optional[ScratchpadAllocator] = None,
+) -> None:
+    # Operations are in topological order (guaranteed by GraphLowering).
+    # Core division has already been done.
+    # Stickification has already been done (therefore all ComputedBuffers have FixedTiledLayouts).
+    if not strategy:
+        strategy = DefaultAllocator()
+    strategy.plan_allocation(graph)
